Add a dummy compiler pass which uses dbg.
This is an example of how to add a pass to the compiler and
uses dbg to facilitate visualizing the flow and the ast.

To support visualizing the ast dbg_ast.h was added and also
a test. Currently dbg_ast only supports emitting to a FILE*.
To support dbg_ctx_t* changes to ast.c will be needed.