Allow fine grained debugging.
The basic concept is there is an opaque debug context which allows
multiple implementations and a simple primary interface:

   DBG(dc, bit_idx)
   DBG_LOC(dc)
   dbg_printf(dc, format, ...)

DBG is a macro which will return true if the bit_idx is set and
also invokes the DBG_LOC macro to set location information.

DBG_LOC sets the location information, function name and line number.

dbg_printf is a variadic functions like fprintf but instead of a FILE*
for the destination it has a dbg_ctx_t*. If there is location information
in the dbg_ctx_t it will preceed the other data. The location information
will be erased after printing so it only prints once.


You also need to create/destroy the dbg_ctx_t. Destroying of all contexts
is done using dbg_destroy(dc). To create a context each implementation
provides one or more create methods. Currently there are three
implementations:

   dbg_ctx_create(number_of_bits)
   dbg_ctx_impl_file_create(number_of_bits, file)
   dbg_ctx_impl_mem_create(number_of_bits, dst_size, tmp_buf_size)

dbg_ctx_create creates an array bits without a destination and is
useful if you never need dbg_printf and just want to control some
other type of information.

dbg_ctx_impl_file_create creates the array of bits and prints to the
file using vfprintf.

dbg_ctx_impl_mem_create creates the array of bits and prints to a memory
buffer using vsnprintf.

Example:

   #define DBG_ENABLED 1
   #include "dbg.h"
   #include "dbg_ctx_impl_file.h"

   void example()
   {
      dbg_ctx_t* dc = dbg_ctx_impl_file_create(2, stdout);
      dbg_sb(dc, 0, 1);
      if(DBG(dc, 0)) dbg_printf(dc, "this is printed with location info\n");
      if(DBG(dc, 1)) dbg_printf(dc, "this is not printed\n");
      dbg_printf(dc, "printed without location %s\n", "info");
      dbg_printf(DBG_LOC(dc), "location %s", "info"); dbg_printf(dc, ", then none!\n");
      dbg_destroy(dc);
   }

The output would be:

   example:9    this is printed with location info
   printed without location info
   example:12   location info, then none!