Use gold linker and fix MAKEFLAGS bug
Use Gold linker by adding LLVM_LINKER=gold to MAKE_OPTIONS in
.packages/deb/rules. It looks to use less memory allowing -j6 to complete
in 55min.

Fix bug, do NOT pass $(MAKEFLAGS) to sub-make's

Passing MAKEFLAGS is not necessary because "most" flags are propagated
already see:
https://www.gnu.org/software/make/manual/html_node/Options_002fRecursion.html#Options_002fRecursion .

This doesn't cause a problem when Makefile-lib-llvm is invoked directly
from the command line because MAKELEVEL=0 and  MAKEFLAGS did not have a
"w" in it. But when Makefile-lib-llvm is invoked from .package/deb/rules
it is not the root, MAKELEVEL > 0, so MAKEFLAGS has a "w". Thus, when
passing $(MAKEFLAGS) on the command line to Makefile-ponyc the "w" ends up
to be a Goal and Makefile-ponyc fails.