Fix bug, do NOT pass $(MAKEFLAGS) to sub-make's.
This is not necessary because "most" flags are propagated already see:
https://www.gnu.org/software/make/manual/html_node/Options_002fRecursion.html#Options_002fRecursion .

This didn't cause a problem directly in Makefile-lib-llvm because
Makefile-lib-llvm was invoked as the root makefile, MAKELEVEL=0, so
MAKEFLAGS did not have a "w" in it.

But when Makefile-lib-llvm is invoked from .package/deb/rules it is not
the root, MAKELEVEL > 0, so MAKEFLAGS has a "w" in it and thus, when
passing $(MAKEFLAGS) on the command line to Makefile-ponyc the "w" ends up to be
as a Goal and Makefile-ponyc fails.